* dev/fix-clang-warnings:
  atomic_count is long, but wrapped function returns int
  vcl_atomic_count returns long not int

+++ core/vil/vil_memory_chunk.h
@@ -54,7 +54,7 @@ class vil_memory_chunk
     void unref();
 
     //: Number of objects referring to this data
-    int ref_count() const { return ref_count_; }
+    long ref_count() const { return ref_count_; }
 
     //: Pointer to first element of data
     virtual void* data();
